requirements.txt
absl-py==0.9.0
astor==0.8.1
blinker==1.4
brotlipy==0.7.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
gast==0.2.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
h5py==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy
oauthlib==3.1.0
opencv-contrib-python==4.4.0.42
opt-einsum==3.1.0
pep517==0.6.0
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrealsense2==2.36.0.2038
python-dateutil==2.8.1
pytoml==0.1.21
pyusb==1.0.2
PyYAML==5.3.1
qimage2ndarray==1.8.3
requests
requests-oauthlib==1.3.0
scikit-learn==0.22
scipy
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
urllib3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
